move more to testx_util
