+++ setup.py
@@ -40,16 +40,42 @@ See https://github.com/PyThaiNLP/pythainlp for installation options.
 requirements = [
     "backports.zoneinfo; python_version<'3.9'",
     "requests>=2.22.0",
-    "tzdata; sys_platform == 'win32'"
+    "tzdata; sys_platform == 'win32'",
 ]
 
 extras = {
+    "abbreviation": ["khamyo>=0.2.0"],
     "attacut": ["attacut>=1.0.6"],
     "benchmarks": ["PyYAML>=5.3.1", "numpy>=1.22", "pandas>=0.24"],
+    "coreference_resolution": [
+        "fastcoref>=2.1.5",
+        "spacy>=3.0",
+    ],
+    "dependency_parsing": [
+        "spacy_thai>=0.7.1",
+        "transformers>=4.22.1",
+        "ufal.chu-liu-edmonds>=1.0.2",
+    ],
+    "el": ["multiel>=0.5"],
+    "esupar": [
+        "esupar>=1.3.8",
+        "numpy",
+        "transformers>=4.22.1",
+    ],
+    "generate": ["fastai<2.0"],
     "icu": ["pyicu>=2.3"],
     "ipa": ["epitran>=1.1"],
     "ml": ["numpy>=1.22", "torch>=1.0.0"],
+    "mt5": ["sentencepiece>=0.1.91", "transformers>=4.6.0"],
+    "nlpo3": ["nlpo3>=1.2.2"],
+    "onnx": ["numpy>=1.22", "onnxruntime>=1.10.0", "sentencepiece>=0.1.91"],
+    "oskut": ["oskut>=1.3"],
+    "sefr_cut": ["sefr_cut>=1.1"],
+    "spacy_thai": ["spacy_thai>=0.7.1"],
+    "spell": ["phunspell>=0.1.6", "spylls>=0.1.5", "symspellpy>=6.7.6"],
     "ssg": ["ssg>=0.0.8"],
+    "textaugment": ["bpemb", "gensim>=4.0.0"],
+    "thai_nner": ["thai_nner"],
     "thai2fit": ["emoji>=0.5.1", "gensim>=4.0.0", "numpy>=1.22"],
     "thai2rom": ["numpy>=1.22", "torch>=1.0.0"],
     "translate": [
@@ -59,47 +85,22 @@ extras = {
         "torch>=1.0.0",
         "transformers>=4.6.0",
     ],
-    "wunsen": ["wunsen>=0.0.1"],
-    "textaugment": ["bpemb", "gensim>=4.0.0"],
-    "wangchanberta": ["sentencepiece>=0.1.91", "transformers>=4.6.0"],
-    "mt5": ["sentencepiece>=0.1.91", "transformers>=4.6.0"],
-    "wtp": ["transformers>=4.6.0", "wtpsplit>=1.0.1"],
-    "wordnet": ["nltk>=3.3"],
-    "generate": ["fastai<2.0"],
-    "sefr_cut": ["sefr_cut>=1.1"],
-    "spell": ["phunspell>=0.1.6", "spylls>=0.1.5", "symspellpy>=6.7.6"],
-    "oskut": ["oskut>=1.3"],
-    "nlpo3": ["nlpo3>=1.2.2"],
-    "onnx": ["numpy>=1.22", "onnxruntime>=1.10.0", "sentencepiece>=0.1.91"],
-    "thai_nner": ["thai_nner"],
-    "esupar": [
-        "esupar>=1.3.8",
-        "numpy",
-        "transformers>=4.22.1",
-    ],
-    "spacy_thai": ["spacy_thai>=0.7.1"],
     "transformers_ud": [
         "transformers>=4.22.1",
         "ufal.chu-liu-edmonds>=1.0.2",
     ],
-    "dependency_parsing": [
-        "spacy_thai>=0.7.1",
-        "transformers>=4.22.1",
-        "ufal.chu-liu-edmonds>=1.0.2",
-    ],
-    "coreference_resolution": [
-        "fastcoref>=2.1.5",
-        "spacy>=3.0",
-    ],
-    "word_approximation": ["panphon>=0.20.0"],
+    "wangchanberta": ["sentencepiece>=0.1.91", "transformers>=4.6.0"],
     "wangchanglm": [
         "pandas>=0.24",
         "sentencepiece>=0.1.91",
         "transformers>=4.6.0",
     ],
+    "word_approximation": ["panphon>=0.20.0"],
+    "wordnet": ["nltk>=3.3"],
     "wsd": ["sentence-transformers>=2.2.2"],
-    "el": ["multiel>=0.5"],
-    "abbreviation": ["khamyo>=0.2.0"],
+    "wtp": ["transformers>=4.6.0", "wtpsplit>=1.0.1"],
+    "wunsen": ["wunsen>=0.0.1"],
+    "compact": ["numpy>=1.22", "pyicu>=2.3", "python-crfsuite>=0.9.7"],
     "full": [
         "PyYAML>=5.3.1",
         "attacut>=1.0.4",
