+++ tests/test_util.py
@@ -65,7 +65,6 @@ from pythainlp.util import (
     words_to_num,
 )
 from pythainlp.util.morse import morse_decode, morse_encode
-from pythainlp.util.spell_words import spell_word
 
 
 class TestUtilPackage(unittest.TestCase):
@@ -803,14 +802,6 @@ class TestUtilPackage(unittest.TestCase):
     def test_tis620_to_utf8(self):
         self.assertEqual(
             tis620_to_utf8("¡ÃÐ·ÃÇ§ÍØµÊÒË¡ÃÃÁ"), "กระทรวงอุตสาหกรรม"
-        )
-
-    def test_spell_word(self):
-        self.assertEqual(spell_word("เสือ"), ["สอ", "เอือ", "เสือ"])
-        self.assertEqual(spell_word("เสื้อ"), ["สอ", "เอือ", "ไม้โท", "เสื้อ"])
-        self.assertEqual(spell_word("คน"), ["คอ", "นอ", "คน"])
-        self.assertEqual(
-            spell_word("คนดี"), ["คอ", "นอ", "คน", "ดอ", "อี", "ดี", "คนดี"]
         )
 
     def test_remove_repeat_consonants(self):
